Now using redis for server side session. This allowes more people to fit in said session so the player isnt limited to only 25 friends at a time. Added top score and rewrote win/lose text. Removed dummy image and unneeded files, renamed files for clarity.
