+++ .gitignore
@@ -64,6 +64,7 @@ db.sqlite3-journal
 # Flask stuff:
 instance/
 .webassets-cache
+flask_session/*
 
 # Scrapy stuff:
 .scrapy
