+++ static/sytles.css
@@ -169,3 +169,7 @@ button:hover {
     grid-area: 4 / 4 / 5 / 5;
     place-self: center center;
 }
+.topSkoor {
+    grid-area: 4 / 1 / 4 / 1;
+    place-self: center center;
+}
