+++ templates/game.html
@@ -32,7 +32,8 @@
             </div>
         </div>
         <h3 class="skoor">Skoor: 0</h3>
+        <h3 class="topSkoor">Top Skoor: 0</h3>
     </div>
-    <script src="./static/ohboy.js" defer></script>
+    <script src="./static/middleman.js" defer></script>
 </body>
 </html>
