+++ templates/win.html
@@ -14,11 +14,14 @@
   <body>
     <div class="container">
       <div class="landing">
-        <h1 class>Võitsid</h1>
-        <p>Tubli! Sa võitsid ja said {{score}} punkti!</p>
-        <p>See on maksimum!</p>
+        <h1 class>Võitsid?</h1>
+        <p>
+          Tubli! Suutsid kõik {{score}} sõpra panna väljamõeldud tähtsuse järgi
+          ritta.
+        </p>
+        <p>See on maksimum, mis on võimalik. Oled nüüd uhke v?</p>
         <form>
-          <button formaction="/game" class="alusta">Mängi uuesti?</button>
+          <button formaction="/game" class="alusta">Tahad veel mängida?</button>
         </form>
       </div>
       <img src="./static/logo.png" />
